pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-harness"
version = "0.1.0"
description = "Secure, governed agent harness for enterprise software"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Orchestration
    "langgraph>=0.4",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-community>=0.3",

    # Observability & evaluation
    "langfuse>=2.0",
    "deepeval>=2.0",

    # Identity & crypto
    "cryptography>=44.0",
    "pyjwt>=2.9",

    # Audit & storage
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "aiosqlite>=0.20",

    # Runtime
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.28",
    "structlog>=24.0",

    # Policy engine
    "pyyaml>=6.0",
    "jsonschema>=4.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_harness"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
